conditionals loops updated

+++ 03-conditionals-loops/VolumeOfPyramid.java
@@ -5,11 +5,11 @@ public class VolumeOfPyramid {
         float l, b, h;
         Scanner input = new Scanner(System.in);
         System.out.print("Enter the length of the pyramid base: ");
-        l = input.nextInt();
+        l = input.nextFloat();
         System.out.print("Enter the breadth of the pyramid base: ");
-        b = input.nextInt();
+        b = input.nextFloat();
         System.out.print("Enter the height of the pyramid: ");
-        h = input.nextInt();
+        h = input.nextFloat();
         float volume = (l*b*h)/3;
         System.out.println("Volume Of Pyramid is: "+volume);
     }
